Controller

+++ src/main/java/pl/coderslab/controller/BookController.java
@@ -1,6 +1,5 @@
 package pl.coderslab.controller;
 
-
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
@@ -15,83 +14,96 @@ import java.util.List;
 import java.util.Objects;
 
 
-@Controller
+@RestController
 @RequestMapping(produces = "application/json; charset=UTF-8")
 public class BookController {
+    private static final String BOOK_NOT_FOUND_MESSAGE="Nie znaleziono książki o podanym Id";
     private final BookDao bookDao;
     private final PublisherDao publisherDao;
     private final AuthorDao authorDao;
 //    private static final System.Logger logger = (System.Logger) LoggerFactory.getLogger(BookController.class);
 
-    public BookController(BookDao bookDao, PublisherDao publisherDao, AuthorDao authorDao) {// wstrzykujemy Dao
+    public BookController(final BookDao bookDao, final PublisherDao publisherDao, final AuthorDao authorDao) {// wstrzykujemy Dao
 
         this.bookDao = bookDao;
         this.publisherDao = publisherDao;
         this.authorDao = authorDao;
     }
-
-
-    @RequestMapping("/book/add")
+//POST
+    @PostMapping("/book/add")
     @ResponseBody
-    public String newBook() {
+    public String newBook(
+            @RequestParam("title") String title,
+            @RequestParam("rating") int rating,
+            @RequestParam("description") String description) {
+
         Book book = new Book();
 
-        book.setTitle("Tytuł1");
-        book.setRating(3);
-        book.setDescription("Description1");
+        book.setTitle(title);
+        book.setRating(rating);
+        book.setDescription(description);
 
 //zadanie1
         Publisher publisher = new Publisher();
         publisher.setName("publisher1");
-        publisherDao.createPublisher(publisher);
+        publisherDao.create(publisher);
         book.setPublisher(publisher);
 //zadanie2
         List<Author> authors = new ArrayList<>();
-        Author author1 = authorDao.findAuthorById(1);
-        Author author2 = authorDao.findAuthorById(2);
+        Author author1 = authorDao.read(1);
+        Author author2 = authorDao.read(2);
         authors.add(author1);
         authors.add(author2);
         book.setAuthors(authors);
 
-        bookDao.saveBook(book);//wywołanie metody utworzonej w klasie BookDao
+        bookDao.create(book);//wywołanie metody utworzonej w klasie BookDao
         // do obiektu pobieramy nadany podczas zapisu identyfikator id
-        return "Id nowej ksiażki: " + book.getId() + " Tytuł nowej książki to: " + book.getTitle() + " rating nowej książki to: "+book.getRating()+ " Opis nowej książki to: "+book.getDescription();
+        return "Id nowej książki: " + book.getId() + " Tytuł nowej książki to: " + book.getTitle() + " rating nowej książki to: "+book.getRating()+ " Opis nowej książki to: "+book.getDescription();
+    }
+    //GET
+    @GetMapping(path = "/book/get")
+    @ResponseBody
+    public String getBook(@RequestParam("id") long id) {
+        Book book = bookDao.read(id);
+        return Objects.nonNull(book)? book.toString(): BOOK_NOT_FOUND_MESSAGE;
     }
 
-    @RequestMapping("/book/all")
-    @ResponseBody
-    public String findAllBooks() {
-        List<Book> books = bookDao.findAll();
-//        books.forEach(b -> logger.info(b.toString()));
-        return Objects.nonNull(books) ? books.toString() : "Brak książek";
-    }
     @GetMapping(path = "/book/rating", produces = "text/plain;charset=UTF-8")
     String getRatingList(@RequestParam("rating") int rating) {
         List<Book> books = bookDao.getRatingList(rating);
         return books.toString();
     }
-    @RequestMapping("/book/{id}/{title}")
-    public String updateBook(@PathVariable long id, @PathVariable String title) {
-        Book book = bookDao.findById(id);
-        book.setTitle(title);
+    @PutMapping("/book}")
+    public String updateBook(@RequestParam("id") long id, @RequestParam("title") String newTitle) {
+        Book book = bookDao.read(id);
+        if(Objects.isNull(book)){
+            return BOOK_NOT_FOUND_MESSAGE;
+        }
+        book.setTitle(newTitle);
         bookDao.update(book);
         return book.toString();
     }
-
-    @RequestMapping("/book/get/{id}")
+    @GetMapping(path = "/book/all")
     @ResponseBody
-    public String getBook(@PathVariable long id) {
-        Book book = bookDao.findById(id);
-        return book.toString();
+    public String findAllBooks() {
+        List<Book> books = bookDao.findAll();
+//        books.forEach(b -> logger.info(b.toString()));
+        return Objects.nonNull(books) ? books.toString() : BOOK_NOT_FOUND_MESSAGE;
     }
-
-    @RequestMapping("book/delete/{id}")
+    //DELETE /book/delete?id=1
+    @DeleteMapping("/book/delete")
     @ResponseBody
-    public String deleteBook(@PathVariable long id) {
-        Book book = bookDao.findById(id);
+    public String deleteBook(@RequestParam("id") long id) {
+        Book book = bookDao.read(id);
+        if(Objects.isNull(book)){
+            return BOOK_NOT_FOUND_MESSAGE;
+        }
         bookDao.delete(book);
         return " Book deleted";
     }
+
+
+
 }
 //Utwórz kontroler o nazwie BookController.
 //Utwórz akcje kontrolera, które będą wykonywać następujące operacje:
